backend/requirements/base.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
django-environ==0.10.0
psycopg2==2.9.6
drf-spectacular==0.26.4
requests==2.31.0
selenium==4.10.0
Pillow==10.0.0